requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.3.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# CORS & Middleware
python-dotenv==1.0.0

# Graph algorithms (para tus algoritmos de grafos)
networkx==3.2.1
numpy==1.26.3
scipy==1.11.4

# Data processing
pandas==2.1.4
openpyxl==3.1.2

# Utilities
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
